Move parallel code to separate package
